+++ packages/morak-ui/src/components/Input/Textarea.css.ts
@@ -3,6 +3,11 @@ import { style } from '@vanilla-extract/css';
 import { input } from './index.css';
 
 export * from './index.css';
+
+export const fullWidth = style({
+  flexGrow: 1,
+});
+
 export const textarea = style([
   input,
   {
@@ -10,3 +15,12 @@ export const textarea = style([
     width: '100%',
   },
 ]);
+
+export const titleWrapper = style({
+  display: 'flex',
+  justifyContent: 'space-between',
+  alignItems: 'center',
+  width: '100%',
+  minHeight: '1.2rem',
+  marginBottom: '0.4rem',
+});
